pyproject.toml
[tool.poetry]
name = "denoising_diffusion"
version = "0.1.0"
description = ""
authors = ["Rajiv M N"]

[tool.poetry.dependencies]
python = "^3.12"
black = "^21.9b0"
flake8 = "^3.9.2"
ipython = "^7.27.0"
isort = "^5.9.3"
jupyterlab = "^3.1.12"
loguru = "^0.5.3"
matplotlib = "^3.4.3"
mkdocs = "^1.2.2"
notebook = "^6.4.3"
numpy = "^1.21.2"
pandas = "^1.3.3"
python-dotenv = "^0.19.0"
scikit-learn = "1.6.1"
tqdm = "^4.62.2"
torch = "*"
torchvision = "*"
torchaudio = "*"
einops = "*"
kagglehub = "*"
ema_pytorch = "*"


[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"